gnuclad.cpp:     fix: empty basename with directory parser gnuclad-cladogram.cpp:     fix: double to int assignment in optimise_strictOverlaps() src/generator/svg.cpp:     fix: validxml() chars 91-96     fix: three double to int assignments src/generator/png.cpp:     added inkscape instructions

+++ src/gnuclad-cladogram.cpp
@@ -910,7 +910,7 @@ bool Cladogram::optimise_strictOverlaps(Node * n, int oldOffset, int sign,
   else
     slope = 0.6;
 
-  int addPX = offsetPX/slope * (n->offset - oldOffset        +1);
+  int addPX = int(offsetPX/slope * (n->offset - oldOffset        +1));
   int stopPX = datePX(n->stop, this) + addPX;
   int startPX = datePX(n->start, this);
